🏗️ Architect: Modularized event DTO calculations in apts/events/
- Extracted calculation helpers and rules from apts/events/event.py into apts/events/calculations/event.py
- Re-exported functions in apts/events/calculations/__init__.py and apts/events/event.py for full public API compatibility
- Refactored Event.__init__ and Event.from_dict to delegate calculation logic to pure helper functions
- Added unit tests in tests/unit/test_events_calculations_event.py
