+++ apts/events/calculations/__init__.py
@@ -4,6 +4,21 @@ from .evaluations import (
     get_duration,
     get_rarity,
 )
+from .event import (
+    CATEGORY_RULES,
+    DirectionData,
+    build_event_description,
+    build_event_title,
+    extract_coordinates,
+    extract_event_objects,
+    format_angular_separation,
+    get_direction_data,
+    get_event_category,
+    get_sky_brightness,
+    get_step_by_step_guide,
+    parse_event_datetime,
+    resolve_event_topocentric_position,
+)
 from .lunar import (
     calculate_lunar_eclipses,
     calculate_lunar_features,
@@ -31,6 +46,19 @@ __all__ = [
     "calculate_event_duration",
     "get_rarity",
     "get_duration",
+    "CATEGORY_RULES",
+    "DirectionData",
+    "get_sky_brightness",
+    "get_direction_data",
+    "get_event_category",
+    "get_step_by_step_guide",
+    "build_event_description",
+    "parse_event_datetime",
+    "extract_event_objects",
+    "build_event_title",
+    "format_angular_separation",
+    "extract_coordinates",
+    "resolve_event_topocentric_position",
     "calculate_lunar_eclipses",
     "calculate_lunar_features",
     "calculate_lunar_occultations",
